+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [v0.4.0] - 2026-03-20
+
+### Added
+- Manually dispatching events without an SSE relay via
+  `Monalisten.dispatch_event` (useful for testing)
+
+### Changed
+- Synced with githubkit 0.15.0
+- Bumped lower bound for githubkit to 0.15.0
+
+
 ## [v0.3.2] - 2025-12-12
 
 ### Changed
@@ -84,3 +95,4 @@ Initial release 🎉
 [v0.3.0]: https://github.com/trag1c/monalisten/compare/v0.2.0...v0.3.0
 [v0.3.1]: https://github.com/trag1c/monalisten/compare/v0.3.0...v0.3.1
 [v0.3.2]: https://github.com/trag1c/monalisten/compare/v0.3.1...v0.3.2
+[v0.4.0]: https://github.com/trag1c/monalisten/compare/v0.3.2...v0.4.0